pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-demo"
version = "0.1.0"
description = "Minimal LangGraph demo with CLI, Cursor rules/commands/skills, and BugBot demo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
langgraph-demo = "langgraph_demo.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["langgraph_demo*"]
